requirements.txt
ccxt==4.3.77
pandas==2.2.2
python-dotenv==1.0.1
